--- src/app/components/ToDoCheck.tsx
@@ -1,76 +1,0 @@
-import styles from "./todocheck.module.css";
-
-interface ToDoCheckProps {
-  checked: boolean;
-}
-
-export default function ToDoCheck(props: ToDoCheckProps) {
-  if (props.checked === true) {
-    return (
-      <>
-        {/* checked */}
-        {/* checkbox */}
-        <label>
-          {/* <input
-              className=
-              {styles.checkbox}
-              type="checkbox"
-              value="Make a to do list app"
-            /> */}
-          <svg
-            xmlns="http://www.w3.org/2000/svg"
-            width="24"
-            height="24"
-            viewBox="0 0 24 24"
-            fill="none"
-          >
-            <circle cx="12" cy="12" r="12" fill="#26BD81" />
-            <path
-              fill-rule="evenodd"
-              clip-rule="evenodd"
-              d="M18.2419 8.13964C18.7228 8.58922 18.7188 9.31447 18.2331 9.75954L11.5662 15.8686C11.3311 16.0841 11.0126 16.2034 10.6819 16.2C10.3511 16.1966 10.0355 16.0707 9.80566 15.8505L5.74755 11.9629C5.27265 11.508 5.28612 10.7828 5.77762 10.3432C6.26912 9.90366 7.05255 9.91612 7.52745 10.3711L10.7152 13.4249L16.4919 8.13147C16.9776 7.6864 17.7611 7.69006 18.2419 8.13964Z"
-              fill="white"
-            />
-          </svg>
-        </label>
-        {/* todo label */}
-        <label className="w-full font-pretendard font-medium text-[16px] text-[#929294] leadign-[24px] tracking-tight line-through">
-          Make a to do list app
-        </label>
-      </>
-    );
-  } else {
-    return (
-      <>
-        {/* unchecked */}
-        {/* checkbox */}
-        <label>
-          {/* <input
-              className={styles.checkbox}
-              type="checkbox"
-              value="Learn JavaScript project"
-              /> */}
-          <svg
-            xmlns="http://www.w3.org/2000/svg"
-            width="24"
-            height="24"
-            viewBox="0 0 24 24"
-            fill="none"
-          >
-            <circle
-              cx="12"
-              cy="12"
-              r="11.4"
-              stroke="#CCCCCE"
-              stroke-width="1.2"
-            />
-          </svg>
-        </label>
-        {/* todo label */}
-        <label className="w-full font-pretendard font-medium text-[16px] text-[#323233] leadign-[24px] tracking-tight">
-          Learn JavaScript project
-        </label>
-      </>
-    );
-  }
-}
