+++ src/app/components/login/createNewAccout.tsx
@@ -1,0 +1,11 @@
+export default function CreateNewAccount() {
+  return (
+    <>
+      <div className="w-[327px] h-[52px] mt-[141px] px-[1px] pt-[1px] pb-[2px] m-auto bg-[#DBDBDE] rounded-[12px]">
+        <button className="w-full h-[49px] px-[24px] py-[12px] rounded-[12px] bg-[#FFFFFF] font-pretendard font-semibold text-[16px] text-[#26BD81]">
+          Create new account
+        </button>
+      </div>
+    </>
+  );
+}
