+++ src/app/components/login/findPassword.tsx
@@ -1,0 +1,11 @@
+export default function FindPassword() {
+  return (
+    <>
+      <div className="w-[327px] mt-[17px] flex justify-center m-auto">
+        <button className="font-pretendard font-semibold text-[#929294] text-[13px]">
+          Forget password?
+        </button>
+      </div>
+    </>
+  );
+}
