+++ src/app/components/signup/signupTitle.tsx
@@ -1,0 +1,10 @@
+export default function SignUpTitle() {
+  return (
+    <>
+      <div className="mx-[1.5rem] mt-[0.5rem] font-pretendard font-semibold text-[1.5rem] leading-[2.375rem]">
+        Start your <br />
+        Phone number or Email
+      </div>
+    </>
+  );
+}
