+++ src/app/components/todolist/ToDoTitle.tsx
@@ -1,0 +1,10 @@
+export default function ToDoTitle() {
+  return (
+    <>
+      {/* To-Do List Title */}
+      <div className="ml-[24px] mt-[32px] font-pretendard font-semibold text-[24px] text-[#323233]">
+        To-Do List
+      </div>
+    </>
+  );
+}
