+++ src/app/todolist/page.tsx
@@ -1,48 +1,15 @@
-import ToDoAdd from "../components/ToDoAdd";
-import ToDoProgress from "../components/ToDoPrgress";
-import ToDoDelete from "../components/ToDoDelete";
-import ToDoCheck from "../components/ToDoCheck";
+import ToDoTitle from "../components/todolist/ToDoTitle";
+import ToDoAdd from "../components/todolist/ToDoAdd";
+import ToDoItems from "../components/todolist/ToDoItems";
+import ToDoProgress from "../components/todolist/ToDoPrgress";
 
 export default function ToDoList() {
   return (
     <>
-      {/* To-Do List Title */}
-      <div className="ml-[24px] mt-[32px] font-pretendard font-semibold text-[24px] text-[#323233]">
-        To-Do List
-      </div>
-      {/* To-Do Add */}
-      <ToDoAdd />
-      {/* To-Do List, Check, Delete */}
-      <div className="h-[408px] mx-[24px] mt-[32px]">
-        {/* To-Do Item :unchecked */}
-        <div className="w-full h-[24px] flex items-center gap-[10px]">
-          {/* checkbox */}
-          <ToDoCheck checked={false} />
-          {/* delete */}
-          <ToDoDelete />
-        </div>
-        {/* line after added */}
-        <div className="my-[20px]">
-          <svg
-            xmlns="http://www.w3.org/2000/svg"
-            width="327"
-            height="2"
-            viewBox="0 0 327 2"
-            fill="none"
-          >
-            <path d="M0 1H327" stroke="#F3F3F5" />
-          </svg>
-        </div>
-        {/* To-Do Item :checked, added */}
-        <div className="w-full h-[24px] flex items-center gap-[10px]">
-          {/* checkbox */}
-          <ToDoCheck checked={true} />
-          {/* delete */}
-          <ToDoDelete />
-        </div>
-      </div>
-      {/* Tasks Done */}
-      <ToDoProgress />
+      <ToDoTitle /> {/* 타이틀 컴포넌트 */}
+      <ToDoAdd /> {/* ToDo 추가 컴포넌트 */}
+      <ToDoItems /> {/* ToDo 목록 컴포넌트 */}
+      <ToDoProgress /> {/* ToDo 진행률 컴포넌트 */}
     </>
   );
 }
